% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffval-package.R
\docType{package}
\name{diffval-package}
\alias{diffval}
\alias{diffval-package}
\title{diffval: Using the DiffVal Index to Classify Vegetation Data}
\description{
This package contains functions to explore vegetation data aiming at obtaining classifications based on differential taxa.
}
\author{
\strong{Maintainer}: Tiago Monteiro-Henriques \email{tiagomonteirohenriques@gmail.com} (\href{https://orcid.org/0000-0002-4206-0699}{ORCID})

}
\keyword{internal}
