% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRDTP_partition_tdv.R
\name{GRDTP_partition_tdv}
\alias{GRDTP_partition_tdv}
\title{Obtain a partition of a phytosociological table using a Greedy-type algorithm (based on TotDiffVal)}
\usage{
GRDTP_partition_tdv(m, k)
}
\arguments{
\item{m}{A \code{matrix}, i.e. a phytosociological table of 0s (absences) and 1s (presences), where rows correspond to taxa and columns correspond to relevés.}

\item{k}{A \code{numeric} giving the number of desired groups.}
}
\value{
A \verb{numeric vector}, which length is the same as the number of columns of m, with numbers from 1 to \code{k}, representing the group to which the respective column was ascribed.
}
\description{
This function obtains a partition of the columns of a given phytosociological matrix, aiming at high values of the TotDiffVal index using a Greedy-type algorithm.
After selecting \code{k} random columns to work as seeds for each of the desired \code{k} groups, the remaining columns are placed in the partition in the following manner: from those unplaced columns, one is selected randomly and, after evaluation, is placed in the group which maximizes the upcoming TotDiffVal. This is repeated until no more columns are unplaced.
}
\details{
This function uses a Greedy-type procedure to obtain a partition of \code{m}.
Given a phytosociological table (\code{m}, with rows corresponding to taxa and columns corresponding to relevés) this function searches for a k-partition (\code{k}, defined by the user) aiming at high values of the TotDiffVal index. Being a simpler method, it is expected to perform faster than \code{\link{GRASP_partition_tdv}}, yet returning worse partitions in terms of TotDiffVal.
}
\author{
Jorge Orestes Cerdeira and Tiago Monteiro-Henriques. E-mail: \email{tiagomonteirohenriques@gmail.com}.
}
