% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUROBI_k2_optim_tdv.R
\name{GUROBI_k2_optim_tdv}
\alias{GUROBI_k2_optim_tdv}
\title{TotDiffVal (or TotDiffVal1) optimization using GUROBI}
\usage{
GUROBI_k2_optim_tdv(
  m,
  formulation = c("t-independent", "t-dependent"),
  index = "TotDiffVal1",
  TimeLimit = Inf
)
}
\arguments{
\item{m}{A \code{matrix}, i.e. a phytosociological table of 0s (absences) and 1s (presences), where rows correspond to taxa and columns correspond to relevés.}

\item{formulation}{A \code{character},  of integer numbers with the initial partition of the relevés (i.e. a vector with values from 1 to k, with length equal to the number of columns of m, ascribing each relevé to one of the k groups). By default, "random", generates a random initial partition.}

\item{index}{A \code{character}, selecting which index to optimize "TotDiffVal", or "TotDiffVal1". #STILL TODO}

\item{TimeLimit}{A \code{numeric} ("double") with the time limit (in seconds) to be passed as a parameter to GUROBI, Defaults to \code{Inf}, but see Details.}
}
\value{
A \code{list} with the following components:

\describe{
\item{par}{A \code{vector} with the 2-partition that maximizes TotDiffVal (or TotDiffVal1). A global maximum.}
\item{objval}{A \code{numeric} with the maximum TotDiffVal (or TotDiffVal1) found.}
}
}
\description{
This function finds the partition of the columns of a given matrix that maximizes the TotDiffVal (or TotDiffVal1) index.
}
\details{
Given a phytosociological table \code{m} (rows corresponding to taxa and columns corresponding to relevés) this function finds a 2-partition (a partition in two groups) that maximizes TotDiffVal (or TotDiffVal1) index (see http://home.isa.utl.pt/~tmh/), i.e. finds, using the GUROBI otpimizer (see \code{\link[gurobi]{gurobi}}). This partition is a global maximum of TotDiffVal (or TotDiffVal1) for any 2-partitions of the dataset.

For medium-sized matrices the computation time might became prohibitive, thus the use of a time limit (\code{TimeLimit}) is very advisable.
}
\author{
Jorge Orestes Cerdeira and Tiago Monteiro-Henriques. E-mail: \email{tiagomonteirohenriques@gmail.com}.
}
