% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_tdv_HillClimb.R
\name{optim_tdv_HillClimb}
\alias{optim_tdv_HillClimb}
\title{Total Differential Value optimization using Hill-climbing algorithms}
\usage{
optim_tdv_HillClimb(
  m.bin,
  p.initial = "random",
  k,
  n.runs = 1,
  n.sol = 1,
  maxit = 10,
  min.g.size = 1,
  stoch.first = FALSE,
  stoch.neigh.size = 1,
  stoch.maxit = 100,
  full.output = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{m.bin}{A \code{matrix}, i.e. a phytosociological table of 0s (absences) and 1s (presences), where rows correspond to taxa and columns correspond to relevés.}

\item{p.initial}{A \code{vector} or a \code{character}. A \code{vector} of integer numbers with the initial partition of the relevés (i.e. a vector with values from 1 to k, with length equal to the number of columns of \code{m.bin}, ascribing each relevé to one of the k groups). By default, \code{p.initial = "random"}, generates a random initial partition.}

\item{k}{A \code{numeric}, giving the number of desired groups.}

\item{n.runs}{A \code{numeric}, giving the number of runs to perform.}

\item{n.sol}{A \code{numeric}, giving the number of best solutions to keep in the final output. Defaults to 1.}

\item{maxit}{A \code{numeric} giving the number of iterations of the Hill-climbing optimization.}

\item{min.g.size}{A \code{numeric} The minimum number of relevés that a group can contain (must be 1 or higher).}

\item{stoch.first}{A \code{logical}. \code{FALSE} (the default), performs only Hill-climbing on the 1-neighbours; \code{TRUE} first, performs a Stochastic Hill-climbing on \code{n}-neighbours (\code{n} is definded by the parameter \code{stoch.neigh.size}), and only after runs the Hill-climbing search on the 1-neighbours; see description above.}

\item{stoch.neigh.size}{A \code{numeric}, giving the size (\code{n}) of the \code{n}-neighbour for the Stochastic Hill-climbing; only used if \code{stoch.first} = \code{TRUE}. Defaults to 1.}

\item{stoch.maxit}{A \code{numeric}, giving the number of iterations of the Hill-climbing optimization; only used if \code{stoch.first} = \code{TRUE}. Defaults to 100.}

\item{full.output}{A \code{logical}. If \code{FALSE} (the default) the best \code{n.sol} partitions and respective indices are returned. If \code{TRUE} (only available for \code{n.sols = 1}) the output will also contain information on the optimization steps (see below).}

\item{verbose}{A \code{logical}. If \code{FALSE} nothing is printed during the runs. If \code{TRUE}, after each run, the run number is printed as well as and indication if the found partition is a 1-neighbour local maximum.}
}
\value{
If \code{full.output = FALSE}, a \code{list} with (at most) \code{n.sol} best solutions (equivalent solutions are removed). Each best solution is also
a \code{list} with the following components:

\describe{
\item{local_maximum}{A \code{logical} indicating if \code{par} is a 1-neighbour local maximum.}
\item{par}{A \code{vector} with the partition of highest TDV obtained by the Hill-climbing algorithm(s).}
\item{tdv}{A \code{numeric} with the TDV of \code{par}.}
}

If \code{full.output = TRUE}, a \code{list} with just one component (one run only), containing also a list with the following components:

\describe{
\item{res.stoch}{A \code{matrix} with the iteration number (of the Stochastic Hill-climbing phase), the maximum TDV found until that iteration, and the higher TDV among all 1-neighbours.}
\item{par.stoch}{A \code{vector} with the best partition found in the Stochastic Hill-climbing phase.}
\item{tdv.stoch}{A \code{numeric} showing the maximum TDV found in the Stochastic Hill-climbing phase (if selected).}
\item{res}{A \code{matrix} with the iteration number (of the Hill-climbing), the maximum TDV found until that iteration, and the higher TDV among all 1-neighbours.}
\item{local_maximum}{A \code{logical} indicating if \code{par} is a 1-neighbour local maximum.}
\item{par}{A \code{vector} with the partition of highest TDV obtained by the Hill-climbing algorithm(s).}
\item{tdv}{A \code{numeric} with the TDV of \code{par}.}
}
}
\description{
This function searches for partitions of the columns of a given matrix, optimizing the Total Differential Value (TDV).
}
\details{
Given a phytosociological table (\code{m.bin}, rows corresponding to taxa and columns corresponding to relevés) this function searches for
a k-partition (\code{k} defined by the user) optimizing TDV, i.e. searches, using a Hill-climbing algorithm, for patterns of differential taxa by
rearranging the relevés into k groups.

Optimization can start from a random partition (\code{p.ini = "random"}), or from a given partition (\code{p.ini}, defined by the user or produced by any
clustering method, or even a manual classification).

Each iteration searches for a TDV improvement screening all 1-neighbours, until the given number of maximum iterations (\code{maxit}) is reached. A
1-neighbour of a given partition is another partition obtained by changing 1 relevé (of the original partition) to a different group. A n-neighbour
is obtained, equivalently, ascribing n relevés to different groups.

Optionally, a faster search (Stochastic Hill-climbing) can be performed in a first step (\code{stoch.first} = \code{TRUE}), consisting on searching for
TDV improvements, by randomly selecting n-neighbours (n defined by the user with the parameter \code{stoch.neigh.size}), accepting that neighour partition
as a better solution if it improves TDV. This is repeated until a given number of maximum iterations (\code{stoch.maxit}) is reached. Stochastic Hill-climbing
might be helpful for big tables (where the simple screening of all 1-neighbours might be too time consuming).

Several runs of optim_tdv_HillClimb (i.e. multiple starts) should be tried out, as several local maxima are usually present and the Hill-climbing
algorithm converges easily to local maxima.

Trimming your table by a 'constancy' range or using the result of other cluster methodologies as input, might help finding interesting partitions.
Specially after trimming the table by a 'constancy' range, getting a random initial partition with TDV greater than zero might be very unlikely; on
such cases using a initial partition from \code{\link{partition_tdv_GRASP}} or \code{\link{partition_tdv_GRDTP}} (or even the result of other clustering
strategies) as input might be useful.
}
\examples{

#getting the Taxus baccata forests data set
data(taxus_bin)

#removing taxa occurring in only one relevé in order to
#try reproducing the example in the original article of the data set
taxus_bin_wmt <- taxus_bin[rowSums(taxus_bin) > 1,]

#obtaining a partition that maximizes TDV using the Stochastic Hill-climbing
#and the Hill-climbing algorithms

result <- optim_tdv_HillClimb(taxus_bin_wmt, k = 3, n.runs = 7, n.sol = 2,
  min.g.size = 3, stoch.first = TRUE, stoch.maxit = 500, verbose = TRUE)

#inspect the result
#the highest TDV found in the runs
result[[1]]$tdv
#if result[[1]]$tdv is 0.1958471 you are probably reproducing the three
#groups (Estrela, Gerês and Galicia) from the original article; if not
#try again the optim_tdv_HillClimb function (maybe increasing n.runs)

#plot the sorted (or tabulated) phytosociological table
tabul1 <- tabulation(taxus_bin_wmt, result[[1]]$par, rownames(taxus_bin_wmt), "normal")

#plot the sorted (or tabulated) phytosociological table, also including
#taxa with occurring just once in the matrix
tabul2 <- tabulation(taxus_bin, result[[1]]$par, rownames(taxus_bin), "normal")

}
\author{
Tiago Monteiro-Henriques. E-mail: \email{tmh.dev@icloud.com}.
}
