% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_taxa.R
\name{select_taxa}
\alias{select_taxa}
\title{Select taxa by a 'constancy' range}
\usage{
select_taxa(m, const = c(0, 100), min.pres = NULL)
}
\arguments{
\item{m}{A \code{matrix}, i.e. a phytosociological table of 0s (absences) and 1s (presences), where rows correspond to taxa and columns correspond to relevés.}

\item{const}{A \code{vector} with two numeric values, which will be used as minimum (the first value) and maximum (the second value) 'constancy' values to trim table \code{m}.}

\item{min.pres}{A \code{numeric} integer giving, optionally, the minimum number presences of each taxon (in all relevés). If provided, it overrides the 'constancy' range.}
}
\value{
A \code{list} with the following components:

\describe{
\item{remo.rel}{The indices of the removed relevé(s).}
\item{remo.sp}{The indices of the removed taxon(taxa).}
\item{m.select}{The trimmed table.}
}
}
\description{
Trims a phytosociological table using the given 'constancy' range.
}
\details{
The function accepts a phytosociological table (\code{m}), and minimum and maximum 'constancy' values (\code{const}). It returns a trimmed phytosociological table, removing taxa outside the given 'constancy' range and removing relevés that become empty. Optionally, a minimum number of presences (in relevés) might be used (\code{min.pres}) instead of the 'constancy' range.
}
\author{
Tiago Monteiro-Henriques. E-mail: \email{tiagomonteirohenriques@gmail.com}.
}
