% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffval-package.R
\docType{package}
\name{diffval-package}
\alias{diffval}
\alias{diffval-package}
\title{diffval: Vegetation Patterns}
\description{
Find, visualize and explore patterns of differential taxa in vegetation data (namely in a phytosociological table), using the Differential Value (DiffVal). Patterns are searched through mathematical optimization algorithms. Ultimately, Total Differential Value (TDV) optimization aims at obtaining classifications of vegetation data based on differential taxa, as in the traditional geobotanical approach. The Gurobi optimizer, as well as the R package 'gurobi', can be installed from \url{https://www.gurobi.com/products/gurobi-optimizer/}. The useful vignette Gurobi Installation Guide, from package 'prioritizr', can be found here: \url{https://prioritizr.net/articles/gurobi_installation_guide.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/point-veg/diffval}
  \item Report bugs at \url{https://gitlab.com/point-veg/diffval/-/issues}
}

}
\author{
\strong{Maintainer}: Tiago Monteiro-Henriques \email{tmh.dev@icloud.com} (\href{https://orcid.org/0000-0002-4206-0699}{ORCID})

Authors:
\itemize{
  \item Jorge Orestes Cerdeira (\href{https://orcid.org/0000-0002-3814-7660}{ORCID})
}

Other contributors:
\itemize{
  \item Fundação para a Ciência e a Tecnologia, Portugal (<https://www.fct.pt/>) [funder]
}

}
\keyword{internal}
