% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffval-package.R
\docType{package}
\name{diffval-package}
\alias{diffval}
\alias{diffval-package}
\title{diffval: Vegetation Patterns}
\description{
Find, visualize and explore patterns of 'differential taxa' in vegetation data (namely in a phytosociological table), using the Differential Value (DiffVal). Patterns are searched through mathematical optimization algorithms. Ultimately, TDV optimization aims at obtaining classifications of vegetation data based on 'differential taxa', as in the traditional phytosociological (or geobotanical) approach. The 'Gurobi' optimizer, as well as the R package 'gurobi', can be installed from <https://www.gurobi.com/products/gurobi-optimizer/>. The vignette "Gurobi Installation Guide", from package 'prioritizr', is useful and can be found here: <https://cran.r-project.org/web/packages/prioritizr/vignettes/gurobi_installation.html>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/point-veg/diffval}
  \item Report bugs at \url{https://gitlab.com/point-veg/diffval/-/issues}
}

}
\author{
\strong{Maintainer}: Tiago Monteiro-Henriques \email{tmh.dev@icloud.com} (\href{https://orcid.org/0000-0002-4206-0699}{ORCID})

Authors:
\itemize{
  \item Jorge Orestes Cerdeira (\href{https://orcid.org/0000-0002-3814-7660}{ORCID})
}

Other contributors:
\itemize{
  \item Fundação para a Ciência e a Tecnologia, Portugal [funder]
}

}
\keyword{internal}
