% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffval-package.R
\docType{package}
\name{diffval-package}
\alias{diffval}
\alias{diffval-package}
\title{diffval: Using the Differential Value to Classify Vegetation Data}
\description{
Obtaining classifications of vegetation data based on differential taxa, as in the traditional phytosociological approach.
}
\author{
\strong{Maintainer}: Tiago Monteiro-Henriques \email{tiagomonteirohenriques@gmail.com} (\href{https://orcid.org/0000-0002-4206-0699}{ORCID})

Authors:
\itemize{
  \item Jorge Orestes Cerdeira \email{tiagomonteirohenriques@gmail.com} (\href{https://orcid.org/0000-0002-4206-0699}{ORCID})
}

Other contributors:
\itemize{
  \item Fundação para a Ciência e a Tecnologia, Portugal \email{tiagomonteirohenriques@gmail.com} (\href{https://orcid.org/0000-0002-4206-0699}{ORCID}) [funder]
}

}
\keyword{internal}
