% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalent_partition.R
\name{equivalent_partition}
\alias{equivalent_partition}
\title{Are the two k-partitions equivalent?}
\usage{
equivalent_partition(p1, p2)
}
\arguments{
\item{p1}{A \code{vector} of integers representing a k-partition (taking values from 1 to k), of the same length of \code{p2}.}

\item{p2}{A \code{vector} of integers representing a k-partition (taking values from 1 to k), of the same length of \code{p1}.}
}
\value{
\code{TRUE} if the k-partitions are equivalent; \code{FALSE} otherwise.
}
\description{
Checks if two k-partitions are equivalent.
}
\details{
The function checks if the two given k-partitions are equivalent, i.e. if their groupings are the same (useful
in the case that the partitions have a divergent group numbering).
}
\author{
Tiago Monteiro-Henriques. E-mail: \email{tiagomonteirohenriques@gmail.com}.
}
