% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_tabulation.R
\name{explore_tabulation}
\alias{explore_tabulation}
\title{Interactively explore a tabulation of a phytosociologic matrix.}
\usage{
explore_tabulation(x)
}
\arguments{
\item{x}{A \code{list} as returned by the \code{\link{tabulation}} function.}
}
\value{
Returns invisibly, although it plots taxa names on the console upon the user click.
}
\description{
This function plots an interactive image of a tabulation.
}
\details{
The function explore.tabulation accepts an object returned by the \code{\link{tabulation}} function, plotting a condensed image of the respective tabulated matrix, permitting the user to click on the coloured blocks and receive the respective list of taxa names on the console.
}
\author{
Tiago Monteiro-Henriques. E-mail: \email{tiagomonteirohenriques@gmail.com}.
}
