% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUROBI_k2_optim_tdv.R
\name{GUROBI_k2_optim_tdv}
\alias{GUROBI_k2_optim_tdv}
\title{Total Differential Value optimization using GUROBI}
\usage{
GUROBI_k2_optim_tdv(m, formulation = "t-dependent", TimeLimit = 5)
}
\arguments{
\item{m}{A \code{matrix}, i.e. a phytosociological table of 0s (absences) and 1s (presences), where rows correspond to taxa and columns correspond to relevés.}

\item{formulation}{A \code{character} selecting wich formulation to use. Possible values are "t-dependent" (the default) or "t-independent". See Details.}

\item{TimeLimit}{A \code{numeric} ("double") with the time limit (in seconds) to be passed as a parameter to GUROBI, Defaults to 5 seconds, but see Details.}
}
\value{
For \code{formulation = "t-dependent"}, a \code{list} with the following components:

\describe{
\item{status.runs}{A \code{character} with GUROBI output status for all the runs.}
\item{objval}{A \code{numeric} with the maximum TDV found by GUROBI.}
\item{par}{A \code{vector} with the 2-partition corresponding to the the maximum TDV found by GUROBI.}
}

For \code{formulation = "t-independent"}, a \code{list} with the following components:

\describe{
\item{status}{A \code{character} with GUROBI output status.}
\item{objval}{A \code{numeric} with the maximum TDV found by GUROBI.}
\item{par}{A \code{vector} with the 2-partition corresponding to the the maximum TDV found by GUROBI.}
}
}
\description{
Given a phytosociological matrix, this function finds the partition of its columns that maximizes the Total Differential Value (TDV).
}
\details{
Given a phytosociological table \code{m} (rows corresponding to taxa and columns corresponding to relevés) this function finds a 2-partition (a
partition in two groups) that maximizes TDV, using the GUROBI otpimizer (see \code{\link[gurobi]{gurobi}}).
When successful, this partition is a global maximum of TDV for any 2-partitions of the columns on \code{m}.

See \code{\link{tdv}} for an explanation on the Total Differential Value of a phytosociological table.

The function implements two different mixed-integer linear programming formulations of the problem. The formulations differ as one is independent of the size
of the obtained groups (t-independent), while the other formulation fixes the size of the obtained groups (t-dependent). The t-dependent formulation is
implemented to run GUROBI as many times as necessary to cover all possible group sizes; this approach can result in faster total computation time.

For medium-sized matrices the computation time might become already prohibitive, thus the use of a time limit (\code{TimeLimit}) is advisable.
}
\examples{

#getting the Taxus baccata forests data set
data(taxus_bin)

#obtaining the 2-partition that maximizes TDV using the GUROBI solver, by
#mixed-integer linear programming
\dontrun{
#requires the suggested package 'gurobi'
GUROBI_k2_optim_tdv(taxus_bin)
}

}
\author{
Jorge Orestes Cerdeira and Tiago Monteiro-Henriques. E-mail: \email{tiagomonteirohenriques@gmail.com}.
}
