% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulAnne_optim_tdv.R
\name{SimulAnne_optim_tdv}
\alias{SimulAnne_optim_tdv}
\title{TotDiffVal (or TotDiffVal1) optimization using the simulated annealing (and GRASP) algorithm(s)}
\usage{
SimulAnne_optim_tdv(
  m,
  k,
  index = "TotDiffVal",
  p.initial = NULL,
  n.runs = 10,
  n.sol = 5,
  T_inic = 1,
  T_final = 0.001,
  alpha = 0.05,
  n.iter = 1000,
  use.GRASP = TRUE,
  full.output = FALSE,
  consider.zeros = TRUE,
  improve.seeds = TRUE
)
}
\arguments{
\item{m}{A \code{matrix}, i.e. a phytosociological table of 0s (absences) and 1s (presences), where rows correspond to taxa and columns correspond to relevés.}

\item{k}{A \code{numeric} giving the number of desired groups.}

\item{index}{A \code{character} selecting which index is calculated: "TotDiffVal" or "TotDiffVal1".}

\item{p.initial}{A \code{vector} of integer numbers with the partition of the relevés (i.e. a k-partition, consisting in a vector with values from 1 to k, with length equal to the number of columns of m, ascribing each relevé to one of the k groups), to be used as initial partition in the simulated annealing, when \code{use.GRASP = FALSE}.}

\item{n.runs}{A \code{numeric} giving the number of runs (it will define the number of seeds). Defaults to 10.}

\item{n.sol}{A \code{numeric}, giving the number of best solutions to keep in the final output. Defaults to 5.}

\item{T_inic}{A \code{numeric} giving the initial temperature. Defaults to 1.}

\item{T_final}{A \code{numeric} giving the initial temperature. Defaults to 0.001.}

\item{alpha}{A \code{numeric} giving the fraction of temperature drop to be used in the temperature reduction scheme (see Details). Defaults to 0.05.}

\item{n.iter}{A \code{numeric} giving the initial temperature. Defaults to 1000}

\item{use.GRASP}{A \code{logical}. Defaults to \code{TRUE}. IF \code{TRUE}, a GRASP procedure is performed on the seeds and then passed to a new simulated annealing phase. If \code{FALSE} the previously obtained seeds are passed to subsequent phase. If \code{TRUE} only that first phase of simulated annealing is performed returning a collection of possible seeds for further analysis.}

\item{full.output}{A \code{logical}. Defaults to \code{FALSE}. IF \code{TRUE} extra information is presented in the output. See Value..}

\item{consider.zeros}{A `logical. TEMPORARY ARGUMENT TO DO SOME TESTING. ELIMINATE AFTER.}

\item{improve.seeds}{A `logical. TEMPORARY ARGUMENT TO DO SOME TESTING. ELIMINATE AFTER.}
}
\value{
If \code{full.output = FALSE} (the default), a \code{list} with the following component:

\describe{
\item{SANN}{A \code{list} with \code{n.runs} components, each one containing also a \code{list} with two components:
\enumerate{
\item 'par', containing the partition found in that run, corresponding to the highest value of TotDiffVal (or TotDiffVal1);
\item 'tdv', with the value of TotDiffVal (or TotDiffVal1) of the referred partition.}
}
}

If \code{full.output = TRUE}, a \code{list} with the following components (the first two components are only given if \code{use.GRASP = TRUE}):

\describe{
\item{improved.seeds}{A \code{list} with \code{n.runs} components, each one containing also a \code{list} with two components: 'seed', containing the \code{k} columns of \code{m} used as improved seed in that run, and 'partial.tdv', which corresponds to the TotDiffVal (or TotDiffVal1) of only those \code{k} columns.}
\item{GRASP}{A \code{list} with \code{n.runs} components, each one containing also a \code{list} with two components: 'par', containing the partition obtained by GRASP, and 'tdv', which corresponds to the TotDiffVal (or TotDiffVal1) of that partition.}
\item{SANN}{A \code{list} with \code{n.runs} components, each one containing also a \code{list} with six components:
\enumerate{
\item 'par', containing the partition found in that run, corresponding to the highest value of TotDiffVal (or TotDiffVal1);
\item 'tdv', with the value of TotDiffVal (or TotDiffVal1) of the referred partition;
\item 'cur.tdv', a \code{vector} of length \code{n.iter} with the current TotDiffVal (or TotDiffVal1) in each SANN iteration;
\item 'alt.tdv', a \code{vector} of length \code{n.iter} with the alternative TotDiffVal (or TotDiffVal1) in each SANN iteration;
\item 'probability', a \code{vector} of length \code{n.iter} with the probability TotDiffVal (or TotDiffVal1) in each SANN iteration;
\item 'temperature', a \code{vector} of length \code{n.iter} with the temperature TotDiffVal (or TotDiffVal1) in each SANN iteration.
}
}
}
}
\description{
This function searches for partitions of the columns of a given matrix, optimizing the TotDiffVal (or TotDiffVal1) index using simulated annealing (and GRASP) algorithm(s).
}
\details{
This function uses two main algorithms:
\enumerate{
\item A greedy randomized adaptive search procedure (GRASP), which is used to obtain first solutions (partitions of \code{m}) that can then be submitted to the next algorithm.
\item A stochastic global optimization method, simulated annealing (SANN), which is used to search for the global maximum of TotDiffVal.
}

Given a phytosociological table (\code{m}, with rows corresponding to taxa and columns corresponding to relevés) this function searches for a k-partition (\code{k}, defined by the user) optimizing TotDiffVal (or TotDiffVal1) index (see http://home.isa.utl.pt/~tmh/), i.e. searches, using a SANN algorithm (possible working on GRASP solutions), for the global maximum of the selected index (by rearranging the relevés into k groups).

The default procedure (\code{use.GRASP = TRUE}, which ignores \code{p.ini}) performs the following steps:
\enumerate{
\item A first step where: i) an initial set of 'group seeds' is obtained randomly (i.e. \code{n.runs} sets of \code{k} columns of \code{m}); ii) these 'group seeds' are submitted to a first SANN optimization, obtaining sets of 'improved group seeds' (i.e., \code{k} columns of \code{m} with high value of TotDiffVal (or TotDiffVal1)); iii) The 'improved group seeds' are submitted to the GRASP to obtain partitions of the columns of \code{m} matrix.
\item In a second step, the partitions obtained by GRASP are used as initial partitions in a SANN optimization.
}

The temperature reduction scheme (CooS) used in the SANN procedure is...

Argument \code{alpha}must be bounded between 0 and 1.
}
\author{
Jorge Orestes Cerdeira and Tiago Monteiro-Henriques. E-mail: \email{tiagomonteirohenriques@gmail.com}.
}
