% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulation.R
\name{tabulation}
\alias{tabulation}
\title{Rearrange a phytosociologic matrix, showing each differential taxa on top.}
\usage{
tabulation(m, p, taxa.names, plot.im = NULL)
}
\arguments{
\item{m}{A \code{matrix}, i.e. a phytosociological table of 0s (absences) and 1s (presences), where rows correspond to taxa and columns correspond to relevés.}

\item{p}{A \code{vector} of integer numbers with the partition of the relevés (i.e. a k-partition, consisting in a vector with values from 1 to k, with length equal to the number of columns of m, ascribing each relevé to one of the k groups).}

\item{taxa.names}{A \verb{character vector} (with length equal to the number of rows of \code{m}) with the taxa names.}

\item{plot.im}{By default, \code{NULL}, returns without plotting. If \code{plot.im} = "normal", plots an image of the tabulated matrix. If \code{plot.im} = "condensed", plots an image of the tabulated matrix but presenting sets of differential taxa as solid coloured blocks.}
}
\value{
If \code{plot.im} = \code{NULL}, a \code{list} with the following components:

\describe{
\item{taxa.names}{The given \code{taxa.names}}
\item{taxa.ord}{A \code{vector} with the order of the rows/taxa.}
\item{tabulated}{The rearranged/reordered \code{m} \code{matrix}.}
\item{condensed}{The matrix used to create the "condensed" image.}
}

If \code{plot.im} = "normal", it returns the above list and, additionally, plots an image of the tabulated matrix.
If \code{plot.im} = "condensed", it returns the above list and, additionally, plots an image of the tabulated matrix, but presenting the sets of differential taxa as solid coloured blocks.
}
\description{
This function reorders a phytosociological table rows based on, firstly, the number of groups in which a taxon occurs in, and secondly, on the within-group (adjusted) relative frequency. The columns are reordered using the given k-partition \code{p}.
}
\details{
The function accepts a phytosociological table (\code{m}) and a k-partition of its columns (\code{p}), returning the respective TotDiffVal1 index.
}
\author{
Tiago Monteiro-Henriques. E-mail: \email{tiagomonteirohenriques@gmail.com}.
}
