% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdv.R
\name{tdv}
\alias{tdv}
\title{The total differential value (TotDiffVal1) of a phytosociologic table}
\usage{
tdv(m, p, full.output = FALSE)
}
\arguments{
\item{m}{A \code{matrix}, i.e. a phytosociological table of 0s (absences) and 1s (presences), where rows correspond to taxa and columns correspond to relevés.}

\item{p}{A \code{vector} of integer numbers with the partition of the relevés (i.e. a k-partition, consisting in a vector with values from 1 to k, with length equal to the number of columns of m, ascribing each relevé to one of the k groups).}

\item{full.output}{A \code{logical}, determining the amount of information returned by the function (defaults to \code{FALSE}).}
}
\value{
If \code{full.output} = \code{FALSE}, a \code{list} with the following components:

\describe{
\item{arf}{A \code{matrix} with the adjusted relative frequency (a/b/e) of each taxon in each group (see http://home.isa.utl.pt/~tmh/).}
\item{adp}{A \code{matrix} with the adjusted differential proportion (c/d/e) of each taxon in each group (see http://home.isa.utl.pt/~tmh/).}
\item{DiffVal1}{A \code{matrix} with the DiffVal1 of each taxon (see http://home.isa.utl.pt/~tmh/).}
\item{TotDiffVal1}{A \code{numeric} with the TotDiffVal1 of matrix \code{m} given the partition \code{p} (see http://home.isa.utl.pt/~tmh/).}
}

If \code{full.output} = \code{TRUE}, these extra components are added: afg, t1, t2, t3. These are intermediate matrices used in the computation of TotDiffVal1.
}
\description{
Calculates TotDiffVal1 index (see http://home.isa.utl.pt/~tmh/ for the calculation formula of TotDiffVal1).
}
\details{
The function accepts a phytosociological table (\code{m}) and a k-partition of its columns (\code{p}), returning the respective TotDiffVal1 index.
}
\author{
Tiago Monteiro-Henriques. E-mail: \email{tiagomonteirohenriques@gmail.com}.
}
